LANGCHAIN/Langchain project 1/requirements.txt
langchain
python-dotenv
langchain-openai
beautifulsoup4
langchain_community
faiss-cpu
streamlit
langchain_ollama
langchain
langchain_groq
langchain_core
fastapi
uvicorn
langserve
sse_starlette